рефакторинг структуры приложения
