+++ ConsoleModule/Program.cs
@@ -14,6 +14,7 @@ namespace ConsoleModule
             //TODO
             //text
 
+
             Function sm = new Function("3*x^2 + 4*x + 5*(x+3)*log2(x)");
             //Tree tree = new Tree(test);
             //Console.WriteLine($"Data root: {tree.DataRoot}");
