+++ QuickMathsBL/Data structure/Node.cs
