+++ QuickMathsBL/QuickMathsBL.csproj
@@ -42,12 +42,13 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConsoleHelper.cs" />
+    <Compile Include="Data structure\TreeBuilder.cs" />
     <Compile Include="Derivative.cs" />
     <Compile Include="Function.cs" />
-    <Compile Include="Node.cs" />
+    <Compile Include="Data structure\Node.cs" />
     <Compile Include="SimpleFunction.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Tree.cs" />
+    <Compile Include="Data structure\Tree.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
